Testing audio algos, wip

+++ firmware/src/pattern.h
@@ -11,7 +11,7 @@ void patternInit(void);
 bool patternBootSequence(void);
 bool patternBootSequence2(void);
 void patternBatteryLevel(u8 level);
-void patternAudioLevel(u8 level);
+void patternAudioLevel(u16 level);
 void clearLeds(void);
 patternFunc patternNext(void);
 patternFunc patternPrevious(void);
